% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format.strzip}
\alias{format.strzip}
\title{Encode strzip}
\usage{
\method{format}{strzip}(x, ...)
}
\arguments{
\item{x}{strzip object}

\item{...}{urther arguments passed to or from other methods.}
}
\description{
Encode strzip
}
