% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.strzip}
\alias{print.strzip}
\title{Print a compressed string}
\usage{
\method{print}{strzip}(x, ...)
}
\arguments{
\item{x}{A compressed string object}

\item{...}{additional arguments to print}
}
\description{
Print a compressed string
}
