% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strzip.R
\name{smaz_compress}
\alias{smaz_compress}
\title{Compress a string using SMAZ algorithm}
\usage{
smaz_compress(input)
}
\arguments{
\item{input}{A character vector string to compress.}
}
\value{
A list of raw vectors containing the compressed data.
}
\description{
Compress a string using SMAZ algorithm
}
