% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmaz.R
\name{smaz_decompress}
\alias{smaz_decompress}
\title{Decompress a raw vector using SMAZ algorithm}
\usage{
smaz_decompress(input)
}
\arguments{
\item{input}{A raw vector to decompress.}
}
\value{
A character string containing the decompressed data.
}
\description{
Decompress a raw vector using SMAZ algorithm
}
