% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{str_compress_read}
\alias{str_compress_read}
\title{Decompress a character vector from a file}
\usage{
str_compress_read(file)
}
\arguments{
\item{file}{The filename containing the compressed binary data.}
}
\value{
A decompressed character string.
}
\description{
This function reads compressed binary data from a file and decompresses it using Unishox2.
}
