% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strzip.R
\name{str_compress_uni}
\alias{str_compress_uni}
\title{Compress a vector of strings using Unishox2}
\usage{
str_compress_uni(x)
}
\arguments{
\item{x}{A character vector of strings to be compressed.}
}
\value{
A character vector of compressed strings.
}
\description{
This function compresses each string in a vector using the Unishox2 compression algorithm.
}
\examples{
str_compress_uni(c("Hello", "world", "Unishox2"))

}
