% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{str_compress_write}
\alias{str_compress_write}
\title{Compress a vector of strings and write to a file}
\usage{
str_compress_write(x, file)
}
\arguments{
\item{x}{A character vector to be compressed.}

\item{file}{filename where the compressed data will be written.}
}
\value{
The length of the compressed data written to the file (invisibly).
}
\description{
This function compresses a string using Unishox2 and writes the compressed binary data to a file.
}
