% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strzip.R
\name{str_decompress_raw}
\alias{str_decompress_raw}
\title{Decompress a raw vector using SMAZ algorithm}
\usage{
str_decompress_raw(x)
}
\arguments{
\item{x}{A list of raw vectors to decompress.}
}
\value{
A character vector of strings containing the decompressed data.
}
\description{
Decompress a raw vector using SMAZ algorithm
}
\examples{
compressed_lst <- str_compress_raw(c("Hello", "world", "smaz"))
str_decompress_raw(compressed_lst)
}
