% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strzip.R
\name{str_decompress_uni}
\alias{str_decompress_uni}
\title{Decompress a vector of strings using Unishox2}
\usage{
str_decompress_uni(x)
}
\arguments{
\item{x}{A character vector of compressed strings to be decompressed.}
}
\value{
A character vector of decompressed strings.
}
\description{
This function decompresses each string in a compressed character vector that was compressed using Unishox2.
}
\examples{
compressed_vec <- str_compress_uni(c("Hello", "world", "Unishox2"))
str_decompress_uni(compressed_vec)
}
