% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strzip.R
\docType{package}
\name{strzip}
\alias{strzip-package}
\alias{strzip}
\title{strzip}
\description{
This package implements short string compression algorithms.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/schochastics/strzip}
  \item Report bugs at \url{https://github.com/schochastics/strzip/issues}
}

}
\author{
\strong{Maintainer}: David Schoch \email{david@schochastics.net} (\href{https://orcid.org/0000-0003-2952-4812}{ORCID})

Other contributors:
\itemize{
  \item Salvatore Sanfilippo (Author of SMAZ library) [copyright holder]
  \item Arundale Ramanathan (Author of Unishox2 library) [copyright holder]
}

}
