% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{[<-.strzip}
\alias{[<-.strzip}
\title{strzip class assign}
\usage{
\method{[}{strzip}(x, i) <- value
}
\arguments{
\item{x}{character}

\item{i}{element}

\item{value}{value}
}
\description{
strzip class assign
}
