% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.strzip}
\alias{summary.strzip}
\title{Summary for strzip}
\usage{
\method{summary}{strzip}(object, type = "mean", ...)
}
\arguments{
\item{object}{strzip object}

\item{type}{either "mean" for average compression rate, "all" for all
individual rates, "summary" for min/max/average rates}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
Summary for strzip
}
