% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strzip.R
\name{unishox2_compress}
\alias{unishox2_compress}
\title{Compress a single string using Unishox2}
\usage{
unishox2_compress(x)
}
\arguments{
\item{x}{A character string to be compressed.}
}
\value{
A compressed character string.
}
\description{
This function compresses a single string using the Unishox2 compression algorithm.
}
\examples{
compressed <- unishox2_compress("Hello, world!")
print(compressed)
}
