% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strzip.R
\name{unishox2_decompress}
\alias{unishox2_decompress}
\title{Decompress a single string using Unishox2}
\usage{
unishox2_decompress(x)
}
\arguments{
\item{x}{A compressed character string to be decompressed.}
}
\value{
A decompressed character string.
}
\description{
This function decompresses a single compressed string that was compressed using Unishox2.
}
\examples{
compressed <- unishox2_compress("Hello, world!")
decompressed <- unishox2_decompress(compressed)
print(decompressed)
}
