% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strzip.R
\name{unishox2_decompress_vector}
\alias{unishox2_decompress_vector}
\title{Decompress a vector of strings using Unishox2}
\usage{
unishox2_decompress_vector(x)
}
\arguments{
\item{x}{A character vector of compressed strings to be decompressed.}
}
\value{
A character vector of decompressed strings.
}
\description{
This function decompresses each string in a compressed character vector that was compressed using Unishox2.
}
\examples{
compressed_vec <- unishox2_compress_vector(c("Hello", "world", "Unishox2"))
decompressed_vec <- unishox2_decompress_vector(compressed_vec)
print(decompressed_vec)
}
