% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmaz.R
\docType{package}
\name{Rsmaz}
\alias{Rsmaz-package}
\alias{Rsmaz}
\title{SMAZ}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
