% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strzip.R
\name{smaz_decompress}
\alias{smaz_decompress}
\title{Decompress a raw vector using SMAZ algorithm}
\usage{
smaz_decompress(x)
}
\arguments{
\item{x}{A list of raw vectors to decompress.}
}
\value{
A character vector of strings containing the decompressed data.
}
\description{
Decompress a raw vector using SMAZ algorithm
}
\examples{
unishox2_compress(c("Hello", "world", "Unishox2"))
}
