% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strzip.R
\docType{package}
\name{strzip}
\alias{strzip-package}
\alias{strzip}
\title{strzip}
\description{
This package wraps the SMAZ C library. Smaz is a simple compression library suitable for compressing very short strings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/schochastics/Rsmaz}
  \item Report bugs at \url{https://github.com/schochastics/Rsmaz/issues}
}

}
\author{
\strong{Maintainer}: David Schoch \email{david@schochastics.net} (\href{https://orcid.org/0000-0003-2952-4812}{ORCID})

Other contributors:
\itemize{
  \item Salvatore Sanfilippo (Author of SMAZ library) [copyright holder]
}

}
